requirements.txt
python-dotenv
langchain
openai
python-dateutil
slack-sdk 
slack-bolt 
flask
langchain-openai
langchain-google-genai
python-dotenv
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
langchain_community
uvicorn[standard]
pytz
msal
flask-session
apscheduler
flask-talisman
psycopg2
gunicorn
uvicorn
